+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [0.8.0] - 2024-08-08
+### Added
+- Network Security Group (NSG) for AKS vnet
+- `network.tf` to contain network resources for better organization
+
 ## [0.7.0] - 2024-08-07
 ### Added
 - User node pool resource for running workloads